requirements.txt
fastapi
uvicorn[standard]
cryptography
sqlalchemy
psycopg2-binary
python-dotenv
